OvmfPkg ResetSystemLib: Fix VS build error
Contributed-under: TianoCore Contribution Agreement 1.0

git-svn-id: https://svn.code.sf.net/p/edk2/code/trunk/edk2@14557 6f19259b-4bc3-4df7-8a09-765794883524

+++ OvmfPkg/Library/ResetSystemLib/ResetSystemLib.c
@@ -27,7 +27,7 @@ AcpiPmControl (
 {
   ASSERT (SuspendType < 6);
 
-  IoBitFieldWrite16  (PcdGet16 (PcdAcpiPmBaseAddress) + 4, 10, 13, SuspendType);
+  IoBitFieldWrite16  (PcdGet16 (PcdAcpiPmBaseAddress) + 4, 10, 13, (UINT16) SuspendType);
   IoOr16 (PcdGet16 (PcdAcpiPmBaseAddress) + 4, BIT13);
   CpuDeadLoop ();
 }
